+++ frontend/src/metabase/querying/notebook/components/ClauseStep/ClauseStep.tsx
@@ -14,7 +14,7 @@ import {
   NotebookCell,
   NotebookCellAdd,
   NotebookCellItem,
-} from "../../components/NotebookCell";
+} from "../NotebookCell";
 
 import { ClausePopover } from "./ClausePopover";
